# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "OpenAPI definition",
    "version": "v0"
  },
  "servers": [
    {
      "url": "http://localhost:8081",
      "description": "Generated server url"
    }
  ],
  "paths": {
...
    "/kyc/api/organization/save": {
      "post": {
        "tags": [
          "kyc-organization-controller"
        ],
        "operationId": "saveOrganization",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/KycOrganizationDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request – invalid input or path variable"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "KycOrganizationDto": {
        "type": "object",
        "properties": {
          "orgId": {
            "type": "string"
          },
          "orgName": {
            "type": "string"
          },
          "spocEmail": {
            "type": "string"
          },
          "orgLogo": {
            "type": "string"
          },
          "spocName": {
            "type": "string"
          },
          "spocMobileNumber": {
            "type": "string"
          }
        }
      },
...
    }
  }
}